+++ src/main/java/com/wallhack/currencyexchange/model/ExchangeRateDTO.java
@@ -2,4 +2,4 @@ package com.wallhack.currencyexchange.model;
 
 import java.math.BigDecimal;
 
-public record ExchangeRateDTO(long id, int baseCurrency, int targetCurrency, BigDecimal rate) {}
+public record ExchangeRateDTO(long id, long baseCurrency, long targetCurrency, BigDecimal rate) {}
